backwellC/backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary
pandas
openpyxl
django-filter